requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
cryptography==41.0.8
tqdm==4.66.1
pytest==7.4.3
pytest-flask==1.3.0